correct docs re: content-type in GSP's
The contentType directive has no effect when sending a
GSP with the html method. It is required when using the body
method, but both text/html and text/xhtml work to make the email
render as HTML.

I tested version 0.4 for reference: the content-type of the GSP
in the email affected the content-type of the GSP rendered by the
controller, so in that version contentType had to be text/html.