Fixed assertion check to allow headers to be empty
The assertion for the `headers` DLS field checks to see if the map is not empty, but then throws the error "headers cannot be null". I believe this assertion is in error and it should use `Assert.notNull()` instead. This allows the passing of empty header lists and in addition produces a correct error message.

Note that I have not tested this change.