fix(deps): set compileOnly dependencies
The `jakarta-servlet-api` was previously included as a transitive dependency,
but it is no longer available through that path. To resolve the issue,
this commit explicitly adds `jakarta-servlet-api` as a `compileOnly` dependency
to ensure the necessary classes are available during compilation without bundling them into the final artifact.

Also change scopes of `groovy` and `grails-core` dependencies to `compileOnly` as these will always be provided to a grails plugin.